refactor: reorganize neovim config following lazy.nvim best practices
- Extract utilities to dedicated modules (utils/keymap.lua, utils/diagnostics.lua)
- Move plugin keymaps to plugin specs (lsp.lua, neogit.lua)
- Remove duplicate LSP configuration (lazy-config/)
- Delete disabled plugins (blink.lua, flash.lua, temporal.lua)
- Clean up commented code in settings.lua and cmp.lua
