+++ neovimlua/.config/nvim/lua/plugins/init.lua
@@ -14,43 +14,37 @@ end
 vim.opt.rtp:prepend(lazypath)
 
 local plugins = {
-	-- require("plugins.gh"), -- disabling for now in favor of neogit and octo
-	-- require("plugins.temporal"),
-	require("plugins.cmp"),
-	-- require("plugins.blink"), -- this was not great, maybe needs more configuration.
-	require("plugins.codecompanion"),
-	require("plugins.colortheme"),
-	require("plugins.comfortable-motion"),
-	require("plugins.comment"),
-	require("plugins.copilot"),
-	require("plugins.dadbod"),
-	-- require("plugins.flash"),
-	require("plugins.fzf"),
-	require("plugins.fzf-lua"),
-	require("plugins.git-signs"), -- I had hoped this wouldn't be needed with neogit but I can't get neogit to show hunks
-	-- require("plugins.lualine"),
-  require('plugins.lazydev'),
-	require("plugins.lualine-simple"),
-	require("plugins.neo-tree"),
-	require("plugins.neogit"),
-	require("plugins.neotest"),
-	require("plugins.octo"),
-	require("plugins.overseer"),
-	require("plugins.pr"),
-	require("plugins.rhubarb"),
-  -- https://www.youtube.com/watch?v=NJDu_53T_4M
-	require("plugins.surround"),
-	require("plugins.tmux-navigator"),
-	require("plugins.treesitter"),
-	require("plugins.trouble"),
+  require("plugins.cmp"),
+  require("plugins.codecompanion"),
+  require("plugins.colortheme"),
+  require("plugins.comfortable-motion"),
+  require("plugins.comment"),
+  require("plugins.copilot"),
+  require("plugins.dadbod"),
+  require("plugins.easymotion"),
+  require("plugins.fzf"),
+  require("plugins.fzf-lua"),
+  require("plugins.git-signs"),
+  require("plugins.lazydev"),
   require("plugins.lsp"),
   require("plugins.lsp_diagnostic_manipulation"),
+  require("plugins.lualine-simple"),
   require("plugins.markdown"),
+  require("plugins.neo-tree"),
+  require("plugins.neogit"),
+  require("plugins.neotest"),
+  require("plugins.octo"),
+  require("plugins.overseer"),
+  require("plugins.pr"),
+  require("plugins.rhubarb"),
   require("plugins.snake"),
   require("plugins.spectre"),
+  require("plugins.surround"),
+  require("plugins.tmux-navigator"),
   require("plugins.transparent"),
+  require("plugins.treesitter"),
+  require("plugins.trouble"),
   require("plugins.which-key"),
-	require("plugins.easymotion"),
 }
 local opts = {}
