--- neovimlua/.config/nvim/lua/plugins/temporal.lua
@@ -1,4 +1,0 @@
-return {
-  name = "temporal-debugger",
-  dir = "~/temporal-debugger"
-}
