+++ neovimlua/.config/nvim/lua/utils/init.lua
@@ -1,0 +1,6 @@
+local M = {}
+
+M.keymap = require("utils.keymap")
+M.diagnostics = require("utils.diagnostics")
+
+return M
